d_last_score: int to float
git-svn-id: https://www.cgran.org/svn/projects/gen2_rfid/trunk@1028 948f8045-528a-dd11-95b9-00e08143843c

+++ rfid/lib/rfid_tag_decoder_f.h
@@ -32,7 +32,7 @@ class rfid_tag_decoder_f : public gr_block
   int                   d_skip_count;
   int                   d_samples_since_reset;
   int                   d_preamble_miss_threshold;
-  int                   d_last_score;
+  float                 d_last_score;
   int                   d_samples_processed;
  
   gr_msg_queue_sptr	d_ctrl_out;  //Pipe control messages to reader block.
